+++ src/main/java/org/giiwa/framework/web/Model.java
@@ -840,7 +840,7 @@ public class Model {
 	final public void set(Beans<? extends Bean> bs, int s, int n) {
 		if (bs != null) {
 			this.set("list", bs);
-			int total = bs.getTotal();
+			int total = (int) bs.getTotal();
 			if (total > 0) {
 				this.set("total", total);
 			}
